+++ src/mocha-bootstrap.ts
@@ -16,12 +16,16 @@ chai.use(function(chai: any, utils: any){
 		new chai.Assertion(report.valid, 'Result should be failure').to.be.false;
 	}
 
-	function expectError(expectedRule: string, expectedMessage: string){
+	function expectError(expectedRule: string, expectedMessage: string|RegExp){
 		const report = utils.flag(this, 'object') as Report;
 		new chai.Assertion(report.results, 'Result should contain an error').to.have.lengthOf(1);
 		const actual = report.results[0].messages[0];
 		new chai.Assertion(actual.ruleId).to.be.equal(expectedRule);
-		new chai.Assertion(actual.message).to.be.equal(expectedMessage);
+		if (expectedMessage instanceof RegExp){
+			new chai.Assertion(actual.message).to.match(expectedMessage);
+		} else {
+			new chai.Assertion(actual.message).to.be.equal(expectedMessage);
+		}
 	}
 
 	chai.Assertion.addProperty('valid', expectValid);