+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features
 
-- new rule `id-pattern` for requiring a specific format.
+- new rules `id-pattern` and `class-pattern` for requiring a specific formats.
 - new rule `no-dup-class` preventing duplicate classes names on the same
   element.